use getmetricMin and getMetricMax to get the AABB of the known OcTree instead of its max dimension
if only a small part of the Octree is actually used and occupied, this will save lots of NarrowPhase collision checks

Note: the testcases for octree all fail as the number of contacts is somehow higher.
This won't be visible in CI as these tests aren't actually run there.
Still they should be fixed of course. I don't understand them unfortunately ...